feat: Add ProjectionError type and validation logic for protocol projection
