+++ work/learnings.md
@@ -294,3 +294,49 @@ without relying on unstable features.
 1. **Phase 1**: Extract tests (immediate compliance improvement)
 2. **Phase 2**: Module restructuring (long-term maintainability)
 3. **Phase 3**: Validation and integration testing
+
+## Task Completion Review (2025-05-24)
+
+### TASKS.md Status Synchronization ✅
+
+**Completed comprehensive review of TASKS.md to align with actual project state:**
+
+#### Key Updates Made
+
+1. **Task 1.1.6c.2 (Advanced Module Restructuring)**: Updated from incomplete to "MOSTLY COMPLETED"
+   - **Verified actual completion**: 4 out of 5 modules successfully restructured under 300-line guideline
+   - **Duality module**: 475 lines → 94 lines ✅ (extracted to helpers.rs, global_impl.rs, local_impl.rs)
+   - **Global module**: 434 lines → 63 lines ✅ (extracted to protocols.rs, implementations.rs)
+   - **Local module**: 448 lines → 79 lines ✅ (extracted to endpoints.rs, implementations.rs)
+   - **Projection module**: Partially completed (336 lines remaining, 36 lines over guideline)
+
+2. **Task 1.1.6c.3 (Validation)**: Updated compliance metrics with current state
+   - Implementation/test ratios now compliant (duality=1.0:1, global=2.1:1, local=1.4:1)
+   - Only 1 module exceeds guidelines vs. previously 4 modules
+
+3. **Task 1.1.6 (Code Quality)**: Marked as "MOSTLY COMPLETED" 
+   - All critical clippy warnings resolved ✅
+   - Default trait implementations completed ✅
+   - Module structure mostly compliant (4/5 modules) ✅
+
+4. **Task 1.1 (Core Implementation)**: Marked as "COMPLETED" with minor cleanup remaining
+
+#### Current Project Status
+
+- **35 tests passing** ✅
+- **Zero clippy warnings** ✅
+- **Code formatting compliant** ✅
+- **Module structure 80% compliant** (4/5 modules under 300 lines)
+
+#### Next Priorities Identified
+
+1. **Optional**: Complete projection module restructuring (36 lines to reduce)
+2. **High Priority**: Task 1.2.3 - Implement Label Preservation and Transformation Logic
+3. **Medium Priority**: Task 2.1 - Develop Unit Tests for Core Types
+
+### Key Learning: Documentation Maintenance
+
+- **Regular TASKS.md synchronization** is critical for accurate project tracking
+- **Line count verification** provides objective completion metrics
+- **Granular task breakdown** enables precise progress tracking
+- **Status markers** should reflect actual implementation state, not planned state
